pyproject.toml
[project]
name = "aipa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.11.0",
    "datasets>=4.3.0",
    "geocoder>=1.38.1",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.3",
    "jinja2>=3.1.6",
    "jupyter>=1.1.1",
    "langchain>=1.0.4",
    "langchain-anthropic>=1.0.1",
    "langchain-huggingface>=1.0.1",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.2",
    "ollama>=0.6.0",
    "openmeteo-requests>=1.7.4",
    "python-dotenv>=1.0.0",
    "pyttsx3>=2.99",
    "requests>=2.32.5",
    "requests-cache>=1.2.1",
    "smolagents[all]>=1.22.0",
    "tokenizers>=0.22.1",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "transformers>=4.57.1",
]
